Initial modifications to support config files to override default command line options for individual files
This is WIP.
The Parser class now checks for a .cr_conf.yaml file in each directory (it doesn't have to be present).
It contains sets of options (a subset of the command line options) for individual files (keyed on file name).
The options supported are:
    license:
    license_file:
    copyright_software:
    copyright_software_description:
    copyright_holders:
    copyright_years:
    word_wrap:
    syntax:
and two config file specific options:
    ext:     - manually specify file type used from the syntax
    include: - whether or not to process this file

To support the ext: option minor changes were made to the Syntax class