Switch to using recursive call for directory descent
Rather than using the ** in Dir.glob now just glob a single directory
at a time, and descend the tree using recursion. This is to allow
better control of inclusion / exclusion of directories when using
.cr_conf.yml files. For example you might add an entry to .cr_conf.yml
so a directory is excluded:
some_dir:
  include: false

or to include a directory which would otherwise be excluded by default
(eg. a directory starting with a '.')
.some_dir:
  include: true