Miscellaneous small changes
Added a check for whether :include: false is set along with
other options for file. A warning is given, because the other
options will be ignored if the file isn't going to be processed.

Added .csh extension to shell type in syntax.yml
Spelling correction for javascript type

Added end of line at end of command_line.rb

Incremented version so I can build a separate gem to standard version