Added --write-all which write all files from input paths (including files which haven't been modified
This option allows creation of a (near) complete copy of the
directory structure of the input paths in the output dir. The files
which have had copyright added/removed are written modified. The
other files and directories (which would otherwise be excluded
from processing) are just copies of the originals.

The reason for this option is to allow better testing of result of
the copyright addition eg. allowing building of source in the copy of
the directory, allowing diff to the original files etc.